Reorganise ActionMode and its size
